Solved the sidebar scroll for mobile users
Solved the most annoying problem for the mobile clients, the sidebar scroll. Yeah, we've experienced this situation when we couldn't invite someone who's at the last end of the sidebar with our mobile browser cause sidebar spills out of the screen when there's a lot of users connected. But, now this is solved by fixing the height of sidebar container to 90% when shown and to 0% when hidden. Ah! I was really grieving to get rid of this shit.