Change verifyNickname to es5 rather than es6.
We're not really using much es6 features, so we might as well make this a normal function so that people with browsers that are old due to not being able to update them (such as apple devices not getting updates the os after end of life). Also replaces the `const` with a `var` since it doesn't really matter if it's modified or not and that is an es6 feature was well.
This might fix https://github.com/hack-chat/main/issues/33 or at least the issues 'CheekyChops' having. I didn't test this code, but it's such a simple change.